fix: Add function domain input parameter
Part of the S3 key for a Lambda function's code is the domain to which
that function belongs (usually a service name like "CustomerService").
There is not an obvious way to "derive" the value, so we are adding an
another workflow input so that consuming workflows can pass this value
in.

See our ADR for more details:

https://github.com/invitation-homes/technology-decisions/blob/main/0064-lambda-naming-conventions.md

jira: PE-4